Fixed bug which allowed i to be incremented past the end of the array in various functions: pwm_map(), tacho_map(), get_sensor_duty(). (#56)
Fixed bug which allowed i to be incremented past the end of the array in various functions: pwm_map(), tacho_map(), get_sensor_duty(). This can cause fan speeds to drop to minimum when temperature of a sensor exceeds the highest value in the temperature map. For example, setting CONF:SENSOR1:TEMPMAP 25,20,50,100 would result in 0% fan speed when temperature exceeds 50C.